SCC-2158 Add handling for no schema cases
This adds a simple conditional to handle the case when an avro schema name is not provided to the kinesis client. This is handled in the invocation already but this removes a bug that causes the initializer to fail because a schema is not provided.

One test is added to cover this condition as well