NOREF Output UTF-8 string from Avro encoding
When not base64 encoding Avro records, the method should return a UTF-8 encoded string. Otherwise other modules such as logging and JSON cannot handle the ASCII string that is usually returned.

This enforces this by setting the encoding on the StringIO object that generates the returned string from the encode method. This should make it safe to use in all contexts.

If base64 encoding the result the string is still in ASCII, though this is safe as a base64 string can be trusted to contain only well-formed/parseable characters